move last remaining sipi message to project messages
