+++ integration/src/test/scala/org/knora/webapi/responders/admin/AssetPermissionsResponderSpec.scala
@@ -8,9 +8,9 @@ package org.knora.webapi.responders.admin
 import org.apache.pekko.testkit.*
 
 import org.knora.webapi.*
+import org.knora.webapi.messages.admin.responder.projectsmessages.PermissionCodeAndProjectRestrictedViewSettings
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectIdentifierADM.ShortcodeIdentifier
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectRestrictedViewSettingsADM
-import org.knora.webapi.messages.admin.responder.sipimessages.*
 import org.knora.webapi.messages.store.triplestoremessages.RdfDataObject
 import org.knora.webapi.routing.UnsafeZioRun
 import org.knora.webapi.sharedtestdata.SharedTestDataADM
