+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/FilesEndpoints.scala
@@ -11,7 +11,7 @@ import sttp.tapir.generic.auto.schemaForCaseClass
 import sttp.tapir.json.zio.jsonBody
 import zio.ZLayer
 
-import org.knora.webapi.messages.admin.responder.sipimessages.PermissionCodeAndProjectRestrictedViewSettings
+import org.knora.webapi.messages.admin.responder.projectsmessages.PermissionCodeAndProjectRestrictedViewSettings
 import org.knora.webapi.slice.admin.api.AdminPathVariables.projectShortcode
 import org.knora.webapi.slice.admin.api.Codecs.TapirCodec.sparqlEncodedString
 import org.knora.webapi.slice.admin.api.FilesPathVar.filename
